tests/exapp_integration/requirements.txt
# SPDX-FileCopyrightText: 2026 Nextcloud GmbH and Nextcloud contributors
# SPDX-License-Identifier: AGPL-3.0-or-later
fastapi
pytest
pytest-playwright
requests
uvicorn
